modal style edit and created separate modal component
