+++ src/components/header/modal.css
@@ -12,44 +12,90 @@
 }
 
 .modal__container {
-  background: #fff;
+  background: rgb(77, 77, 77);
   border-radius: 5px;
-  max-width: 100%;
+  height: 80vh;
   margin: 50px auto;
   padding: 15px;
-  width: 560px;
+  width: 90vh;
+  z-index: 2;
 }
 
 .modal__title {
   font-size: 18px;
   font-weight: 600;
-  margin-bottom: 25px;
-}
-
-#my-object {
-  height: auto;
-  width: 300px;
-}
-
-#header-text {
   color: var(--color-primary);
+  margin-bottom: 1rem;
 }
 
-/* .a {
-  width: max-content !important;
-  display: inline-block !important;
-  color: var(--color-primary) !important;
-  padding: 0.75rem 1.2rem !important;
-  border-radius: 0.4rem !important;
-  cursor: pointer !important;
-  border: 1px solid var(--color-primary) !important;
-  transition: var(--transition) !important;
-  background: var(--color-primary) !important;
-  color: var(--color-bg) !important;
+.modal__cta {
+  display: flex;
+  gap: 0.5rem;
+  justify-content: center;
 }
 
-.a:hover {
-  background: var(--color-white) !important;
-  color: var(--color-bg) !important;
-  border-color: transparent !important;
-} */
+.modal__btn {
+  font-size: 0.8rem;
+  width: max-content;
+  display: inline-block;
+  background: var(--color-primary);
+  color: var(--color-bg);
+  padding: 0.2rem 0.6rem;
+  border-radius: 0.4rem;
+  cursor: pointer;
+  border: 1px solid var(--color-primary);
+  transition: var(--transition);
+}
+
+.modal__exit {
+  font-size: 0.8rem;
+  padding-bottom: 1rem;
+}
+
+.modal__object {
+  /* height: 50vh;
+  width: 80vh;
+  margin-bottom: 1rem; */
+  /* height: 80%;
+  width: 80%;
+  margin: 0 auto;
+  padding: 10px;
+  position: relative;
+  z-index: 3; */
+
+  height: 500px;
+  width: 100%;
+  padding-bottom: 1rem;
+  position: relative;
+}
+
+.modal__object iframe {
+  height: 100%;
+  width: 100%;
+  box-sizing: border-box;
+}
+/* ==================== MEDIA QUERIES (MEDIUM DEVICES) ==================== */
+@media screen and (max-width: 1024px) {
+  .modal__container {
+    width: var(--container-width-md);
+  }
+}
+
+/* ==================== MEDIA QUERIES (SMALL DEVICES) ==================== */
+@media screen and (max-width: 600px) {
+  .modal__container {
+    width: var(--container-width-sm);
+  }
+
+  /* .modal__object iframe {
+    height: 70%;
+    width: 70%;
+  } */
+
+  .modal__btn {
+    font-size: 0.7rem;
+  }
+  .modal__exit {
+    font-size: 0.7rem;
+  }
+}
